pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-screener"
version = "0.1.0"
description = "AI image detection using frequency domain analysis"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "scipy>=1.10.0",
    "PyWavelets>=1.4.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "opencv-python>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
viz = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

